--- components/SplitPane/build.gradle.kts
@@ -1,21 +1,0 @@
-import org.jetbrains.compose.compose
-
-plugins {
-    kotlin("multiplatform")
-    id("org.jetbrains.compose")
-}
-
-kotlin {
-    jvm {
-        withJava()
-    }
-    sourceSets {
-        named("jvmMain") {
-            dependencies {
-                implementation(compose.desktop.currentOs)
-                implementation(project("common"))
-            }
-        }
-    }
-}
-
