--- components/SplitPane/common/build.gradle.kts
@@ -1,27 +1,0 @@
-import org.jetbrains.compose.compose
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-plugins {
-    kotlin("multiplatform")
-    id("org.jetbrains.compose")
-}
-
-kotlin {
-    jvm("desktop")
-
-    sourceSets {
-        named("commonMain") {
-            dependencies {
-                api(compose.runtime)
-                api(compose.foundation)
-                api(compose.material)
-            }
-        }
-        named("desktopMain") {}
-    }
-}
-
-// TODO it seems that argument isn't applied to the common sourceSet. Figure out why
-tasks.withType<KotlinCompile>().configureEach {
-    kotlinOptions.freeCompilerArgs += "-Xopt-in=kotlin.RequiresOptIn"
-}
