--- components/SplitPane/desktop/build.gradle.kts
@@ -1,29 +1,0 @@
-import org.jetbrains.compose.compose
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-plugins {
-    kotlin("multiplatform")
-    id("org.jetbrains.compose")
-}
-
-kotlin {
-    jvm {}
-    sourceSets {
-        named("jvmMain") {
-            dependencies {
-                implementation(compose.desktop.currentOs)
-                implementation(project(":SplitPane:common"))
-            }
-        }
-    }
-}
-
-compose.desktop {
-    application {
-        mainClass = "org.jetbrains.compose.splitpane.demo.MainKt"
-    }
-}
-
-tasks.withType<KotlinCompile>().configureEach {
-    kotlinOptions.freeCompilerArgs += "-Xopt-in=kotlin.RequiresOptIn"
-}
