+++ components/VideoPlayer/library/build.gradle.kts
