+++ components/build.gradle.kts
@@ -1,6 +1,5 @@
 buildscript {
-    // __LATEST_COMPOSE_RELEASE_VERSION__
-    val composeVersion = System.getenv("COMPOSE_RELEASE_VERSION") ?: "0.4.0"
+    val composeVersion = property("compose.version")
 
     repositories {
         google()
@@ -15,10 +14,37 @@ buildscript {
     }
 }
 
-allprojects {
+subprojects {
+    version = findProperty("deploy.version") ?: property("compose.version")!!
+
     repositories {
         google()
         mavenCentral()
         maven("https://maven.pkg.jetbrains.space/public/p/compose/dev")
+    }
+
+    plugins.withId("java") {
+        configureIfExists<JavaPluginExtension> {
+            sourceCompatibility = JavaVersion.VERSION_11
+            targetCompatibility = JavaVersion.VERSION_11
+
+            withJavadocJar()
+            withSourcesJar()
+        }
+    }
+
+    plugins.withId("maven-publish") {
+        configureIfExists<PublishingExtension> {
+            repositories {
+                maven {
+                    name = "ComposeRepo"
+                    setUrl(System.getenv("COMPOSE_REPO_URL"))
+                    credentials {
+                        username = System.getenv("COMPOSE_REPO_USERNAME")
+                        password = System.getenv("COMPOSE_REPO_KEY")
+                    }
+                }
+            }
+        }
     }
 }
