+++ components/gradle.properties
@@ -2,3 +2,6 @@ org.gradle.jvmargs=-Xmx2048m -Dfile.encoding=UTF-8
 android.useAndroidX=true
 android.enableJetifier=true
 kotlin.code.style=official
+
+# __LATEST_COMPOSE_RELEASE_VERSION__
+compose.version=0.4.0
